+++ src/routes/(admin)/admin/noticesadmin/+page.svelte
@@ -1,27 +1,36 @@
 <script lang="ts">
 	import { onMount } from 'svelte';
 
+	import { addDoc, updateDoc, doc, type DocumentReference, deleteDoc } from 'firebase/firestore';
+	import { noticeColRef } from '$lib/firebase/firebaseConfig';
+
 	import {
 		NoticesStore,
 		CollectionType,
 		loadItems,
-		type NoticeForm,
+		createHashableString,
+		type Notice,
 		type CollectionItem,
 	} from '$lib/stores/ObjectStore';
-
-	import { decodeHtml } from '$lib/utils/HTMLfunctions';
+	import { notificationStore, TOAST_DURATION, Messages } from '$lib/stores/notifications';
 
 	import Button from '$lib/components/ui/button/button.svelte';
 	import Icon from '@iconify/svelte';
 	import Label from '$lib/components/ui/label/label.svelte';
 	import Input from '$lib/components/ui/input/input.svelte';
-	import { Timestamp } from 'firebase/firestore';
-
+	import ToastContainer from '$lib/components/ToastContainer.svelte';
+	import ConfirmDialog from '$lib/components/ConfirmDialog.svelte';
+	import ToastEditor from '$lib/components/ToastEditor.svelte';
 	const loadData = async () => {
 		await loadItems(CollectionType.Notices);
+		console.log($NoticesStore);
 	};
 
 	let loading = $state(true);
+	let mode: 'new' | 'update' = $state('new');
+	let showDeleteDialog = $state(false);
+	let deleteID = $state('');
+	let updateId = '';
 
 	onMount(async () => {
 		await loadData();
@@ -29,38 +38,80 @@
 		loading = false;
 	});
 
-	const initialNotice: NoticeForm = { due: '', text: '' };
+	const initialNotice: Notice = { due: '', text: '' };
 
 	let thisNotice = $state(initialNotice);
 
-	const handleCreateNew = () => {
-		console.log('New');
-	};
-
-	const dateToLocalDatetimeString = (date: Date) => {
-		const year = date.getFullYear();
-		const month = String(date.getMonth() + 1).padStart(2, '0');
-		// padStart: fill the string with "0" to a total length of 2 from the left
-		const day = String(date.getDate()).padStart(2, '0');
-		const hours = String(date.getHours()).padStart(2, '0');
-		const minutes = String(date.getMinutes()).padStart(2, '0');
-		return `${year}-${month}-${day}T${hours}:${minutes}`;
+	const convertDate = (date: string) => {
+		return date.slice(0, 10);
 	};
 
 	const handleEdit = (item: CollectionItem) => {
-		const jsDate = item.data.due.toDate();
-		thisNotice.due = dateToLocalDatetimeString(jsDate);
+		mode = 'update';
+		updateId = item.id;
+		thisNotice.due = convertDate(item.data.due);
 		thisNotice.text = item.data.text;
-		console.log('Item: ', item.data.due);
-		console.log(thisNotice);
 	};
 
-	const handleSubmit = () => {
-		console.log('Ich speichere das Item');
+	const handleSubmit = async () => {
+		if (!thisNotice.due || !thisNotice.text) {
+			notificationStore.addToast('warning', 'Please fill out all mandatory fields', TOAST_DURATION);
+			return;
+		}
+		const noticeData = {
+			due: thisNotice.due + 'T23:59:00',
+			text: thisNotice.text,
+		};
+		if (mode === 'new') {
+			try {
+				const res = await addDoc(noticeColRef, noticeData);
+				const newNotices = [...$NoticesStore, { id: res.id, data: noticeData }];
+				NoticesStore.set(newNotices);
+				notificationStore.addToast('success', Messages.SAVESUCCESS, TOAST_DURATION);
+				handleReset();
+			} catch (error) {
+				notificationStore.addToast('error', Messages.SAVERROR, TOAST_DURATION);
+			}
+		} else if (mode === 'update') {
+			try {
+				const docRef = doc(noticeColRef, updateId);
+				await updateDoc(docRef, noticeData);
+				const updatedNotices = $NoticesStore.map((item) => {
+					if (item.id === updateId) {
+						return { id: item.id, data: noticeData };
+					}
+					return item;
+				});
+				NoticesStore.set(updatedNotices);
+				notificationStore.addToast('success', Messages.UPDATESUCCESS, TOAST_DURATION);
+				handleReset();
+			} catch (error) {
+				console.log(error);
+				notificationStore.addToast('error', Messages.UPDATEERROR, TOAST_DURATION);
+			}
+		}
+	};
+
+	const openModal = (id: string) => {
+		deleteID = id;
+		showDeleteDialog = true;
+	};
+
+	const handleDelete = async () => {
+		try {
+			const docRef = doc(noticeColRef, deleteID);
+			await deleteDoc(docRef);
+			notificationStore.addToast('success', Messages.DELETESUCCESS, TOAST_DURATION);
+			const newNotices = $NoticesStore.filter((notice) => notice.id != deleteID);
+			NoticesStore.set(newNotices);
+		} catch (error) {
+			notificationStore.addToast('error', Messages.DELETEERROR, TOAST_DURATION);
+		}
 	};
 
 	const handleReset = () => {
-		console.log('Ich lösche das Formular');
+		thisNotice = initialNotice;
+		mode = 'new';
 	};
 </script>
 
@@ -75,19 +126,20 @@
 	</div>
 
 	<div class="w-full rounded-xl border p-4">
-		<form id="form-container" enctype="multipart/form-data" onsubmit={handleSubmit} onreset={handleReset}>
+		<h2>{mode === 'new' ? 'Create new notice' : 'Edit notice'}</h2>
+		<form id="form-container" enctype="multipart/form-data">
 			<fieldset>
 				<Label for="duedate">Due Date*</Label>
-				<Input type="datetime-local" id="duedate" bind:value={thisNotice.due} required />
+				<Input type="date" id="duedate" bind:value={thisNotice.due} required />
 			</fieldset>
 			<fieldset>
 				<Label for="text">Notice Text*</Label>
 				<Input type="text" id="text" bind:value={thisNotice.text} required />
-				<p class="py-2 text-sm">Use &lt;b&gt;TEXT&lt;/b&gt; to print <b>TEXT</b></p>
+				<p class="py-2 text-sm">Use &lt;b&gt;TEXT&lt;/b&gt; to print <b>TEXT</b>; use &lt;br&gt; to add a new line</p>
 			</fieldset>
 			<div class="flex flex-row items-center justify-center gap-8">
 				<Button variant="secondary" size="lg" onclick={handleReset}>Reset form</Button>
-				<Button variant="primary" size="lg" onclick={handleCreateNew}>Create News</Button>
+				<Button variant="primary" size="lg" onclick={handleSubmit}>Save</Button>
 			</div>
 		</form>
 	</div>
@@ -108,14 +160,14 @@
 					<tbody class="table-row">
 						{#each $NoticesStore as item}
 							<tr>
-								<td class="table-data table-cell">{item.data.due.toDate().toLocaleString('de-DE')}</td>
+								<td class="table-data table-cell">{convertDate(item.data.due)}</td>
 								<td class="table-data table-cell">{@html item.data.text}</td>
 								<td class="table-data table-cell">
 									<div class="flex justify-between gap-2">
 										<Button variant="primary" class="min-w-0" onclick={() => handleEdit(item)}>
 											<Icon icon="fluent-mdl2:edit" class="size-6" />
 										</Button>
-										<Button variant="destructive" title="Delete Item" class="min-w-0"
+										<Button variant="destructive" title="Delete Item" class="min-w-0" onclick={() => openModal(item.id)}
 											><Icon icon="mdi-light:delete" class="size-6" /></Button
 										>
 									</div>
@@ -128,3 +180,16 @@
 		</div>
 	{/if}
 </div>
+
+<ConfirmDialog
+	open={showDeleteDialog}
+	title="Confirm Delete"
+	message="Deleting an item can not be undone. <br \>Do you really want to delete this item?"
+	confirmText="Delete"
+	cancelText="Cancel"
+	confirmVariant="destructive"
+	onConfirm={handleDelete}
+	onCancel={() => (showDeleteDialog = false)}
+/>
+
+<ToastContainer />